smus-cdk/requirements.txt
constructs>=10.4.2
boto3>=1.38.0
aws-cdk-lib>=2.188.0